移动 公共函数
